added searchView to CurrencyFragment.kt

+++ app/src/main/java/com/example/itogovoe/data/sources/local_source/entities/CurrenciesEntity.kt
@@ -7,7 +7,7 @@ import java.time.LocalDateTime
 @Entity
 data class CurrenciesEntity(
     @PrimaryKey val name: String,
-    val value: Double,
+    val value: Float,
     val createdAt: LocalDateTime,
     val updatedAt: LocalDateTime,
     val lastUsedAt: LocalDateTime,
