+++ app/src/main/java/com/example/itogovoe/domain/model/HistoryDtoModel.kt
@@ -5,7 +5,7 @@ import java.time.LocalDateTime
 data class HistoryDtoModel(
     val date: LocalDateTime,
     val currencyNameParent: String,
-    val currencyValueParent: Double,
+    val currencyValueParent: Float,
     val currencyNameChild: String,
-    val currencyValueChild: Double
+    val currencyValueChild: Float
 )
