+++ app/src/main/java/com/example/itogovoe/ui/model/HistoryUiModel.kt
@@ -3,7 +3,7 @@ package com.example.itogovoe.ui.model
 data class HistoryUiModel(
     val date: String,
     val currencyNameParent: String,
-    val currencyValueParent: Double,
+    val currencyValueParent: Float,
     val currencyNameChild: String,
-    val currencyValueChild: Double
+    val currencyValueChild: Float
 )
